chore: re-export types from colocated files
