+++ packages/browser/src/notifier.d.ts
@@ -1,19 +1,2 @@
-import { Client, Config, BugsnagStatic } from '@bugsnag/core'
-
-interface BrowserConfig extends Config {
-  maxEvents?: number
-  collectUserIp?: boolean
-  generateAnonymousId?: boolean
-  trackInlineScripts?: boolean
-}
-
-interface BrowserBugsnagStatic extends BugsnagStatic {
-  start(apiKeyOrOpts: string | BrowserConfig): Client
-  createClient(apiKeyOrOpts: string | BrowserConfig): Client
-}
-
-declare const Bugsnag: BrowserBugsnagStatic
-
-export default Bugsnag
-export * from '@bugsnag/core'
-export { BrowserConfig }
+export { default } from '../types/bugsnag'
+export * from '../types/bugsnag'
