+++ packages/browser/types/bugsnag.d.ts
@@ -1,2 +1,19 @@
+import { Client, Config, BugsnagStatic } from '@bugsnag/core'
 
-export * from '../src/notifier'
+interface BrowserConfig extends Config {
+  maxEvents?: number
+  collectUserIp?: boolean
+  generateAnonymousId?: boolean
+  trackInlineScripts?: boolean
+}
+
+interface BrowserBugsnagStatic extends BugsnagStatic {
+  start(apiKeyOrOpts: string | BrowserConfig): Client
+  createClient(apiKeyOrOpts: string | BrowserConfig): Client
+}
+
+declare const Bugsnag: BrowserBugsnagStatic
+
+export default Bugsnag
+export * from '@bugsnag/core'
+export { BrowserConfig }
