+++ packages/browser/types/global.d.ts
@@ -1,0 +1,4 @@
+import bugsnag from './bugsnag'
+
+export as namespace bugsnag;
+export = bugsnag;
